Conflicts:
	StortSpelprojekt/AI/Heap.h
	StortSpelprojekt/StortSpelprojekt/Game.cpp
	StortSpelprojekt/StortSpelprojekt/JsonStructs.h
	StortSpelprojekt/StortSpelprojekt/ObjectHandler.cpp
	StortSpelprojekt/StortSpelprojekt/StortSpelprojekt.vcxproj
	StortSpelprojekt/StortSpelprojekt/Trap.cpp
	StortSpelprojekt/StortSpelprojekt/Trap.h

+++ StortSpelprojekt/StortSpelprojekt/Unit.cpp
@@ -44,7 +44,7 @@ void Unit::Rotate()
 
 int Unit::GetApproxDistance(AI::Vec2D target) const
 {
-	return _aStar->GetHeuristicDistance(_tilePosition, target);
+	return (int)_aStar->GetHeuristicDistance(_tilePosition, target);
 }
